Refactor into single AccessTokenProxy model
