+++ apps/badgeuser/signals.py
@@ -1,6 +1,5 @@
 import badgrlog
 
-
 badgrlogger = badgrlog.BadgrLogger()
 
 
@@ -20,3 +19,8 @@ def handle_email_created(sender, instance=None, created=False, **kwargs):
     """
     if created:
         instance.user.publish_method('cached_emails')
+
+def delete_refresh_tokens(refresh):
+    def delete(sender, instance, **kwargs):
+        refresh.objects.filter(access_token=instance.pk).delete()
+    return delete
