+++ apps/badgeuser/tests/__init__.py
@@ -1,0 +1,2 @@
+# encoding: utf-8
+from __future__ import unicode_literals
